scripts/requirements.txt
pandas==2.2.3
numpy==1.26.4
scikit-surprise==1.1.4
